proactaslo/requirements.txt
fastapi
uvicorn
asyncio
confluent-kafka
redis
asyncpg
sqlalchemy
torch
prophet
river
prometheus-client
httpx
grafana
pydantic-settings
python-dotenv
apscheduler
aiohttp
